Create page14stronglyEnum.cpp
I wanted to be sure that enumerators is not something we can simply pass cout streams.
So the way to print their values is at this point of the book still unknown. The reader has no clue.
But after page 16 (where the switch is presented ) one possible way is apparent. See page16switchExample.cpp